requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
pytest>=7.4.0
pytest-cov>=4.1.0
roboflow>=1.1.0
albumentations>=1.3.0
pyyaml>=6.0
scikit-learn>=1.3.0
